// src/app/fisher/shared/interface/fisher-options.interface.ts
import { OptionInterface } from '../../../shared/interface/option.interface';

export interface FisherOptionsInterface {
  propertyCategory: OptionInterface[];
  propertySubCategory: OptionInterface[];
  propertyStatus: OptionInterface[];
  country: OptionInterface[];
  motivation: OptionInterface[];
}